requirements.txt
langgraph
langsmith
langchain
langchain_groq
langchain_community
arxiv
wikipedia
python-dotenv
huggingface-hub
